# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Mimi App Backup V1 API
  version: 0.1.0
  description: Same-origin API for the current Lazycat tenant. Every business endpoint requires a valid OIDC session.
paths:
...
  /auth/login:
    get:
      summary: Begin OIDC Authorization Code with PKCE
      parameters:
        - $ref: '#/components/parameters/ReturnTo'
      responses:
        '302': { description: Redirect to Lazycat OIDC }
        '403': { $ref: '#/components/responses/IdentityMismatch' }
...
  /auth/logout:
    post:
      security: [{ sessionCookie: [] }]
      responses:
        '204': { description: Session deleted }
        '401': { $ref: '#/components/responses/SessionRequired' }
  /api/session:
    get:
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Current verified session
          content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } }
        '401': { $ref: '#/components/responses/SessionRequired' }
        '403': { $ref: '#/components/responses/IdentityMismatch' }
  /api/applications:
    get:
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: q
          in: query
          schema: { type: string, maxLength: 200 }
        - name: mode
          in: query
          schema: { type: string, enum: [single, multi] }
        - name: capability_status
          in: query
          schema: { $ref: '#/components/schemas/CapabilityStatus' }
        - name: protection_status
          in: query
          schema: { type: string, enum: [PROTECTED, UNPROTECTED] }
      responses:
        '200':
          description: Persisted current-tenant application instances
          content: { application/json: { schema: { $ref: '#/components/schemas/ApplicationList' } } }
        '400': { $ref: '#/components/responses/ApiError' }
  /api/applications/sync:
    post:
      summary: Queue a coalesced current-tenant catalog sync
      security: [{ sessionCookie: [] }]
      responses:
        '202':
          description: Sync accepted
          content: { application/json: { schema: { $ref: '#/components/schemas/SyncAccepted' } } }
  /api/applications/{appid}:
    get:
      security: [{ sessionCookie: [] }]
      parameters:
        - name: appid
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Logical application with its current-tenant instances
          content: { application/json: { schema: { $ref: '#/components/schemas/ApplicationDetail' } } }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
  /api/instances/{deployId}:
    get:
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/DeployId'
      responses:
        '200':
          description: Current-tenant application instance
          content: { application/json: { schema: { $ref: '#/components/schemas/ApplicationInstance' } } }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
  /api/instances/{deployId}/probe:
    post:
      summary: Queue a coalesced catalog refresh containing this instance's probe
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/DeployId'
      responses:
        '202':
          description: Probe refresh accepted
          content: { application/json: { schema: { $ref: '#/components/schemas/SyncAccepted' } } }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
  /api/instances/{deployId}/backup-scope:
    get:
      summary: Read safe relative paths for a current-tenant selective backup plan
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/DeployId'
        - name: q
          in: query
          schema: { type: string, maxLength: 200 }
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { description: Relative metadata only, content: { application/json: { schema: { $ref: '#/components/schemas/BackupScopeCatalog' } } } }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
  /api/instances/{deployId}/backup:
    post:
      summary: Queue one bounded manual backup for a current-tenant instance
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/DeployId'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ManualBackupRequest' }
      responses:
        '202':
          description: Manual backup accepted
          content: { application/json: { schema: { $ref: '#/components/schemas/ManualBackupAccepted' } } }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/ApiError' }
        '429': { $ref: '#/components/responses/ApiError' }
  /api/backup-jobs/{jobId}:
    get:
      summary: Read one current-tenant manual backup job
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Manual backup job status
          content: { application/json: { schema: { $ref: '#/components/schemas/BackupJob' } } }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
  /api/backups:
    get:
      summary: List recent completed current-tenant snapshots
      security: [{ sessionCookie: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Snapshot records with safe relative storage metadata
          content: { application/json: { schema: { $ref: '#/components/schemas/SnapshotList' } } }
  /api/backups/{snapshotId}:
    get:
      summary: Read one current-tenant snapshot
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200':
          description: Snapshot metadata
          content: { application/json: { schema: { $ref: '#/components/schemas/Snapshot' } } }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
  /api/backups/{snapshotId}/verify:
    post:
      summary: Perform a fast current-tenant ZIP and manifest verification
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/SnapshotId'
        - name: mode
          in: query
          schema: { type: string, enum: [quick, full], default: quick }
      responses:
        '200':
          description: Verification completed
          content: { application/json: { schema: { $ref: '#/components/schemas/Snapshot' } } }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/ApiError' }
  /api/plans:
    get:
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Current-tenant backup plans
          content: { application/json: { schema: { $ref: '#/components/schemas/PlanList' } } }
    post:
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PlanInput' } } }
      responses:
        '201': { description: Plan created, content: { application/json: { schema: { $ref: '#/components/schemas/BackupPlan' } } } }
        '400': { $ref: '#/components/responses/ApiError' }
        '409': { $ref: '#/components/responses/ApiError' }
  /api/plans/{planId}:
    get:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/PlanId' }]
      responses:
        '200': { description: Current-tenant plan, content: { application/json: { schema: { $ref: '#/components/schemas/BackupPlan' } } } }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
    put:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/PlanId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PlanInput' } } }
      responses:
        '200': { description: Plan updated, content: { application/json: { schema: { $ref: '#/components/schemas/BackupPlan' } } } }
  /api/plans/{planId}/run:
    post:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/PlanId' }]
      responses:
        '202': { description: Batch queued, content: { application/json: { schema: { $ref: '#/components/schemas/PlanRunAccepted' } } } }
  /api/plans/{planId}/pause:
    post:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/PlanId' }]
      responses:
        '200': { description: Plan paused, content: { application/json: { schema: { $ref: '#/components/schemas/BackupPlan' } } } }
  /api/plans/{planId}/resume:
    post:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/PlanId' }]
      responses:
        '200': { description: Plan resumed, content: { application/json: { schema: { $ref: '#/components/schemas/BackupPlan' } } } }
  /api/batches:
    get:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/TaskLimit' }]
      responses:
        '200': { description: Current-tenant batches, content: { application/json: { schema: { $ref: '#/components/schemas/BatchList' } } } }
  /api/batches/{batchId}:
    get:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/BatchId' }]
      responses:
        '200': { description: Current-tenant batch, content: { application/json: { schema: { $ref: '#/components/schemas/BackupBatch' } } } }
  /api/tasks:
    get:
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskLimit' }
        - name: status
          in: query
          schema: { type: string }
        - name: batch_id
          in: query
          schema: { type: string }
      responses:
        '200': { description: Current-tenant tasks, content: { application/json: { schema: { $ref: '#/components/schemas/TaskList' } } } }
  /api/tasks/{taskId}:
    get:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/TaskId' }]
      responses:
        '200': { description: Task and attempts, content: { application/json: { schema: { $ref: '#/components/schemas/TaskDetail' } } } }
  /api/tasks/{taskId}/cancel:
    post:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/TaskId' }]
      responses:
        '200': { description: Cancellation accepted, content: { application/json: { schema: { $ref: '#/components/schemas/BackupTask' } } } }
  /api/tasks/{taskId}/retry:
    post:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/TaskId' }]
      responses:
        '202': { description: Retry queued, content: { application/json: { schema: { $ref: '#/components/schemas/BackupTask' } } } }
  /api/backups/{snapshotId}/files:
    get:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/SnapshotId' }]
      responses:
        '200': { description: ZIP file index, content: { application/json: { schema: { $ref: '#/components/schemas/SnapshotFileList' } } } }
  /api/backups/{snapshotId}/export:
    post:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/SnapshotId' }]
      responses:
        '202': { description: Export completed, content: { application/json: { schema: { $ref: '#/components/schemas/ExportAccepted' } } } }
  /api/storage:
    get:
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Current-tenant storage summary, content: { application/json: { schema: { $ref: '#/components/schemas/StorageSummary' } } } }
  /api/storage/scan:
    post:
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Library scan completed, content: { application/json: { schema: { $ref: '#/components/schemas/StorageSummary' } } } }
  /api/overview:
    get:
      summary: Read the current tenant operational overview
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Current-tenant overview, content: { application/json: { schema: { $ref: '#/components/schemas/Overview' } } } }
  /api/alerts:
    get:
      summary: List current-tenant in-app alerts
      security: [{ sessionCookie: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [OPEN, MUTED, RESOLVED, ALL] }
        - { $ref: '#/components/parameters/TaskLimit' }
      responses:
        '200': { description: Current-tenant alerts, content: { application/json: { schema: { $ref: '#/components/schemas/AlertList' } } } }
  /api/alerts/{alertId}/read:
    post:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AlertId' }]
      responses:
        '200': { description: Alert marked read, content: { application/json: { schema: { $ref: '#/components/schemas/Alert' } } } }
  /api/alerts/{alertId}/resolve:
    post:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AlertId' }]
      responses:
        '200': { description: Alert resolved, content: { application/json: { schema: { $ref: '#/components/schemas/Alert' } } } }
  /api/alerts/{alertId}/mute:
    post:
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AlertId' }]
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/MuteAlertRequest' } } }
      responses:
        '200': { description: Alert temporarily muted, content: { application/json: { schema: { $ref: '#/components/schemas/Alert' } } } }
  /api/settings:
    get:
      summary: Read current-tenant safe defaults
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Current settings, content: { application/json: { schema: { $ref: '#/components/schemas/Settings' } } } }
    put:
      summary: Update current-tenant safe defaults
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Settings' } } }
      responses:
        '200': { description: Saved settings, content: { application/json: { schema: { $ref: '#/components/schemas/Settings' } } } }
        '400': { $ref: '#/components/responses/ApiError' }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: lazycat_backup_session }
  parameters:
    ReturnTo: { name: return_to, in: query, schema: { type: string } }
    Cursor: { name: cursor, in: query, schema: { type: string } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
...
  responses:
    ApiError:
      description: Safe, localizable application error
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    SessionRequired: { $ref: '#/components/responses/ApiError' }
    IdentityMismatch: { $ref: '#/components/responses/ApiError' }
...
  schemas:
    Session:
      type: object
      required: [uid, subject, displayName, role, tenantUid, expiresAt, identityVerified]
      properties:
        uid: { type: string }
        subject: { type: string }
        displayName: { type: string }
        email: { type: string }
        role: { type: string, enum: [NORMAL, ADMIN] }
        tenantUid: { type: string }
        expiresAt: { type: string, format: date-time }
        identityVerified: { type: boolean }
    CapabilityStatus:
      type: string
      enum: [BACKUPABLE, BACKUPABLE_SHARED_WARNING, NO_DATA, UNSUPPORTED_DATABASE, SOURCE_MAPPING_AMBIGUOUS, SOURCE_NOT_READY, SOURCE_PERMISSION_DENIED, SOURCE_INSTANCE_NOT_FOUND, SOURCE_OWNER_MISMATCH, RUNTIME_APPVAR_PROJECTION_NOT_VISIBLE, PROBE_FAILED]
...
    ApplicationList:
      type: object
      required: [items, sync]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ApplicationInstance' } }
        nextCursor: { type: string }
        sync: { $ref: '#/components/schemas/SyncStatus' }
...